requirements.txt
flask>=1.0.2
numpy>=1.15.4
pandas>=0.23.4
pytz>=2018.7
xlsxwriter>=1.1.2
libhxl>=4.11